Switch to emacs-geeknote fork
It seems that geeknote has not been in active development for quite a while. So
a recent fork of it (jeffkowalski/geeknote) is used instead. However, the API
of the fork differed from the original geeknote. So I forked and modified
geeknote.el to address this issue.

This will use my fork of geeknote.el instead of the MELPA one.